Use Resources in HDRP and Post-Processing stack implementation. Fix ImageEffect component. Update Changelog.md. Update demos.
