+++ Changelog.md
@@ -11,18 +11,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Add High-Definition Render Pipeline implementation of Order-Independent Transparency
+- Add Universal Render Pipeline implementation of Order-Independent Transparency
 - Support Vulkan and OpenGL by using SV_SampleIndex for MSAA support
 
 ### Changed
 
-- Remove conditional compilations for PostProcessing package
+- Restructure package for different render pipelines using Assemblies
 - Changed package name to avoid using protected names
 - Update project description and demo scenes
-- Small shader and interface improvements
-
-### Removed
-
-- Remove CameraComponent implementation because of deprecated ImageEffect support
+- Various small shader and interface improvements and updates
 
 ## [2.0.0]
