--- Runtime/PostProcessingStackV2/OitResourcesPostProcessingStack.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 3225ba9ed39fe75478bdd712458cafea
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
