+++ Runtime/Shared/Legacy.meta
@@ -1,8 +1,8 @@
 fileFormatVersion: 2
-guid: b9e3c1a68ffca5617a9309fe66bfe19f
-NativeFormatImporter:
+guid: 2140e8ee1781a7346b787637cbbf9471
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
-  mainObjectFileID: 11400000
   userData: 
   assetBundleName: 
   assetBundleVariant:
