+++ Runtime/Shared/OitResources.cs
