--- Runtime/URP/OitResourcesURP.asset
@@ -1,16 +1,0 @@
-%YAML 1.1
-%TAG !u! tag:unity3d.com,2011:
---- !u!114 &11400000
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: ad4349063661457c8435b97f1b841e47, type: 3}
-  m_Name: OitResourcesURP
-  m_EditorClassIdentifier: 
-  oitFullscreenRender: {fileID: 4800000, guid: f2e2448455bf4f6ba2ec561a0aa09479, type: 3}
-  oitComputeUtils: {fileID: 7200000, guid: 8166d1d02844f514985e551cdad81e49, type: 3}
