--- Samples~/Post Process Demo.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 24196354c92349619c7aebed32b831dc
-timeCreated: 1668630317
