--- Samples~/Post Process Demo/Materials.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 30ba063564de44deaecc6a51cfcda531
-timeCreated: 1668630317
