--- Samples~/Post Process Demo/Materials/BlueTransparent.mat.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 54c0ee33d7b146c5954b9c995ee303c5
-timeCreated: 1668630317
