--- Samples~/Post Process Demo/Materials/GreenTransparent.mat.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 3061938cc5d548ad82c9c1f74168dd5e
-timeCreated: 1668630317
