--- Samples~/Post Process Demo/Materials/RedTransparent.mat.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 5c7fdf4a33ca40359f7fbe017bd92bff
-timeCreated: 1668630317
