+++ Samples~/Post Process Demo/Post-Processing Profile.asset
@@ -1,6 +1,6 @@
 %YAML 1.1
 %TAG !u! tag:unity3d.com,2011:
---- !u!114 &-7186432057460268936
+--- !u!114 &-4516301733053350453
 MonoBehaviour:
   m_ObjectHideFlags: 3
   m_CorrespondingSourceObject: {fileID: 0}
@@ -16,6 +16,9 @@ MonoBehaviour:
   enabled:
     overrideState: 1
     value: 1
+  shaderResources:
+    overrideState: 1
+    value: {fileID: 11400000, guid: 51b7ace4c025007fe889160c4a7641ea, type: 2}
 --- !u!114 &11400000
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -29,4 +32,4 @@ MonoBehaviour:
   m_Name: Post-Processing Profile
   m_EditorClassIdentifier: 
   settings:
-  - {fileID: -7186432057460268936}
+  - {fileID: -4516301733053350453}
