--- Samples~/Post Process Demo/Post-Processing Profile.asset.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 8329b7aeaee84962aaee821b0d0c4b9f
-timeCreated: 1668630317
