--- Samples~/Post Process Demo/Scenes.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 3a53b96cd4a047fba167cb1ee4d855f9
-timeCreated: 1668630317
