--- Samples~/Post Process Demo/Scenes/OrderIndependentTransparency.unity.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 67943e81900c42e48e6ed60c450e907f
-timeCreated: 1668630317
